Enables extra parameters to be passed to the load_from_checkpoint function
# Summary:

When someone trains a model using a GPU and then tries to this model on a machine with CPUs only. They are presented with the following error message

```
RuntimeError: Attempting to deserialize object on a CUDA device but torch.cuda.is_available() is False. If you are running on a CPU-only machine, please use torch.load with map_location=torch.device('cpu') to map your storages to the CPU.
```

If one specifies the following:
```
MODEL_FILENAME_DICT[model_name].load_from_checkpoint(f"{path}/{model}",  map_location=torch.device('cpu'))
```

The error is remedied.

This PR generalizes the error above, by allowing users to pass additional arguments to the `load_from_checkpoint` to avoid such complications.